+++ catalyst/utils/torch.py
@@ -9,7 +9,7 @@ from torch import nn, Tensor
 import torch.backends
 from torch.backends import cudnn
 
-from catalyst.settings import IS_CUDA_AVAILABLE, NUM_CUDA_DEVICES, SETTINGS
+from catalyst.settings import SETTINGS
 from catalyst.typing import (
     Device,
     Model,
@@ -31,46 +31,6 @@ ACTIVATIONS = {  # noqa: WPS407
     nn.LeakyReLU: "leaky_relu",
     nn.ELU: "relu",
 }
-
-from catalyst.engines.torch import DataParallelEngine, DeviceEngine, DistributedDataParallelEngine
-
-if SETTINGS.amp_required:
-    from catalyst.engines.amp import (
-        AMPEngine,
-        DataParallelAMPEngine,
-        DistributedDataParallelAMPEngine,
-    )
-
-if SETTINGS.apex_required:
-    from catalyst.engines.apex import (
-        APEXEngine,
-        DataParallelApexEngine,
-        DistributedDataParallelApexEngine,
-    )
-
-
-def get_engine(fp16: bool = False, apex: bool = False, ddp: bool = False):
-    has_multiple_gpus = NUM_CUDA_DEVICES > 1
-    if not IS_CUDA_AVAILABLE:
-        return DeviceEngine("cpu")
-    else:
-        if fp16 and SETTINGS.amp_required and ddp and has_multiple_gpus:
-            return DistributedDataParallelAMPEngine()
-        elif apex and SETTINGS.apex_required and ddp and has_multiple_gpus:
-            return DistributedDataParallelApexEngine()
-        elif fp16 and has_multiple_gpus:
-            return DataParallelAMPEngine()
-        elif fp16 and NUM_CUDA_DEVICES == 1:
-            return AMPEngine()
-        elif apex and has_multiple_gpus:
-            return DataParallelApexEngine()
-        elif apex and NUM_CUDA_DEVICES == 1:
-            return APEXEngine()
-        elif ddp and has_multiple_gpus:
-            return DistributedDataParallelEngine()
-        elif has_multiple_gpus:
-            return DataParallelEngine()
-    return DeviceEngine("cuda")
 
 
 def _nonlinearity2name(nonlinearity):
